% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/early_zero_weight.R
\name{early_zero_weight}
\alias{early_zero_weight}
\title{Zero early weight for weighted logrank tests}
\usage{
early_zero_weight(x, early_period = 4, fail_rate = NULL)
}
\arguments{
\item{x}{A \code{\link[=counting_process]{counting_process()}}-class \code{tibble} with a counting process dataset.}

\item{early_period}{The initial delay period where weights increase;
after this, weights are constant at the final weight in the delay period.}

\item{fail_rate}{a data frame record the failure rate}
}
\value{
A data frame. The column \code{weight} contains the weights for the
early zero weighted logrank test for the data in \code{x}.
}
\description{
Zero early weight for weighted logrank tests
}
\examples{
library(dplyr)
library(gsDesign2)

# Example 1: unstratified
sim_pw_surv(n = 200) \%>\%
  cut_data_by_event(125) \%>\%
  counting_process(arm = "experimental") \%>\%
  early_zero_weight(early_period = 2) \%>\%
  filter(row_number() \%in\% seq(5, 200, 40))

# Example 2: stratified
n <- 500
# two strata
stratum <- c("Biomarker-positive", "Biomarker-negative")
prevelance_ratio <- c(0.6, 0.4)

# enrollment rate
enroll_rate <- define_enroll_rate(
  stratum = rep(stratum, each = 2),
  duration = c(2, 10, 2, 10),
  rate =  c(c(1, 4) * prevelance_ratio[1], c(1, 4) * prevelance_ratio[2]))
enroll_rate$rate <- enroll_rate$rate * n / sum(enroll_rate$duration * enroll_rate$rate)

# failure rate
med_pos <- 10         # median of the biomarker positive population
med_neg <- 8          # median of the biomarker negative population
hr_pos <- c(1, 0.7)   # hazard ratio of the biomarker positive population
hr_neg <- c(1, 0.8)   # hazard ratio of the biomarker negative population
fail_rate <- define_fail_rate(
  stratum = rep(stratum, each = 2),
  duration = c(3, 1000, 4, 1000),
  fail_rate = c(log(2) / c(med_pos, med_pos, med_neg, med_neg)),
  hr = c(hr_pos, hr_neg),
  dropout_rate = 0.01)

# simulate data
temp <- simfix2simpwsurv(fail_rate)        # transfer the failure rate
set.seed(2023)

sim_pw_surv(
  n = n,                                                            # sample size
  stratum = tibble(stratum = stratum, p = prevelance_ratio),        # stratified design with prevalence ratio of 6:4
  block =  c("control", "control", "experimental", "experimental"), # randomization ratio
  enroll_rate = enroll_rate,                                        # enrollment rate
  fail_rate = temp$fail_rate,                                       # failure rate
  dropout_rate = temp$dropout_rate                                  # dropout rate
  ) \%>\%
  cut_data_by_event(125) \%>\%
  counting_process(arm = "experimental") \%>\%
  early_zero_weight(early_period = 2, fail_rate = fail_rate) \%>\%
  filter(row_number() \%in\% seq(5, 200, 40))
}
\references{
Xu, Z., Zhen, B., Park, Y., & Zhu, B. (2017).
"Designing therapeutic cancer vaccine trials with delayed treatment effect."
\emph{Statistics in medicine}, 36(4), 592-605.
}
