% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ex2delayedEffect.R
\docType{data}
\name{Ex2delayedEffect}
\alias{Ex2delayedEffect}
\title{Time-to-event data example 2 for non-proportional hazards working group}
\format{
Data frame with 4 variables:
\itemize{
\item \code{id}: Sequential numbering of unique identifiers.
\item \code{month}: Time-to-event.
\item \code{event}: 1 for event, 0 for censored.
\item \code{trt}: 1 for experimental, 0 for control.
}
}
\usage{
data(Ex2delayedEffect)
}
\description{
Survival objects reverse-engineered datasets from published Kaplan-Meier
curves.
Individual trials are de-identified since the data are only
approximations of the actual data.
Data are intended to evaluate methods and designs for trials where
non-proportional hazards may be anticipated for outcome data.
}
\examples{
library(survival)

data(Ex2delayedEffect)
km1 <- with(Ex2delayedEffect, survfit(Surv(month, evntd) ~ trt))
km1
plot(km1)
with(subset(Ex2delayedEffect, trt == 1), survfit(Surv(month, evntd) ~ trt))
with(subset(Ex2delayedEffect, trt == 0), survfit(Surv(month, evntd) ~ trt))
}
\references{
Lin, Ray S., Ji Lin, Satrajit Roychoudhury, Keaven M. Anderson, Tianle Hu,
Bo Huang, Larry F Leon, Jason J.Z. Liao, Rong Liu, Xiaodong Luo,
Pralay Mukhopadhyay, Rui Qin, Kay Tatsuoka, Xuejing Wang,
Yang Wang, Jian Zhu, Tai-Tsang Chen, Renee Iacona &
Cross-Pharma Non-proportional Hazards Working Group. 2020.
Alternative analysis methods for time to event endpoints under
nonproportional hazards: A comparative analysis.
\emph{Statistics in Biopharmaceutical Research} 12(2): 187--198.
}
\seealso{
\link{Ex1delayedEffect},
\link{Ex3curewithph},
\link{Ex4belly},
\link{Ex5widening},
\link{Ex6crossing}
}
\keyword{datasets}
