% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ex5widening.R
\docType{data}
\name{Ex5widening}
\alias{Ex5widening}
\title{Time-to-event data example 5 for non-proportional hazards working group}
\format{
Data frame with 4 variables:
\itemize{
\item \code{id}: Sequential numbering of unique identifiers.
\item \code{month}: Time-to-event.
\item \code{event}: 1 for event, 0 for censored.
\item \code{trt}: 1 for experimental, 0 for control.
}
}
\usage{
data(Ex5widening)
}
\description{
Survival objects reverse-engineered datasets from published Kaplan-Meier
curves.
Individual trials are de-identified since the data are only
approximations of the actual data.
Data are intended to evaluate methods and designs for trials where
non-proportional hazards may be anticipated for outcome data.
}
\examples{
library(survival)

data(Ex5widening)
km1 <- with(Ex5widening, survfit(Surv(month, evntd) ~ trt))
km1
plot(km1)
}
\references{
TBD
}
\seealso{
\link{Ex1delayedEffect},
\link{Ex2delayedEffect},
\link{Ex3curewithph},
\link{Ex4belly},
\link{Ex6crossing}
}
\keyword{datasets}
