% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ex6crossing.R
\docType{data}
\name{Ex6crossing}
\alias{Ex6crossing}
\title{Time-to-event data example 6 for non-proportional hazards working group}
\format{
Data frame with 4 variables:
\itemize{
\item \code{id}: Sequential numbering of unique identifiers.
\item \code{month}: Time-to-event.
\item \code{event}: 1 for event, 0 for censored.
\item \code{trt}: 1 for experimental, 0 for control.
}
}
\usage{
data(Ex6crossing)
}
\description{
Survival objects reverse-engineered datasets from published Kaplan-Meier
curves.
Individual trials are de-identified since the data are only
approximations of the actual data.
Data are intended to evaluate methods and designs for trials where
non-proportional hazards may be anticipated for outcome data.
}
\examples{
library(survival)

data(Ex6crossing)
km1 <- with(Ex6crossing, survfit(Surv(month, evntd) ~ trt))
km1
plot(km1)
}
\references{
TBD
}
\seealso{
\link{Ex1delayedEffect},
\link{Ex2delayedEffect},
\link{Ex3curewithph},
\link{Ex4belly},
\link{Ex5widening}
}
\keyword{datasets}
