% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedBlockRand.r
\name{fixedBlockRand}
\alias{fixedBlockRand}
\title{Permuted fixed block randomization}
\usage{
fixedBlockRand(n = 10, block = c(0, 0, 1, 1))
}
\arguments{
\item{n}{sample size to be randomized}

\item{block}{Vector of treatments to be included in each block}
}
\value{
A treatment group sequence (vector) of length `n` with treatments from `block` permuted within
each block having block size equal to the length of `block`
}
\description{
Fixed block randomization. The `block` input should repeat each treatment code the number of
times it is to be included within each block. The final block will be a partial block if `n` is not an
exact multiple of the block length.
}
\examples{
library(dplyr)
# 2:1 randomization with block size 3, treatments "A" and "B"
tibble(x=1:10) \%>\% mutate(Treatment=fixedBlockRand(block=c("A","B","B")))
# Stratified randomization
tibble(Stratum=c(rep("A",10),rep("B",10))) \%>\%
group_by(Stratum) \%>\%
mutate(Treatment=fixedBlockRand())
}
