% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cut_date_by_event.R
\name{get_cut_date_by_event}
\alias{get_cut_date_by_event}
\title{Get date at which an event count is reached}
\usage{
get_cut_date_by_event(x, event)
}
\arguments{
\item{x}{A time-to-event dataset, for example, generated by \code{\link[=sim_pw_surv]{sim_pw_surv()}}.}

\item{event}{Event count at which dataset is to be cut off for analysis.}
}
\value{
A numeric value with the \code{cte} from the input dataset
at which the targeted event count is reached, or if the final event count
is never reached, the final \code{cte} at which an event occurs.
}
\description{
Get date at which an event count is reached
}
