% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCutDateForCount.r
\name{getCutDateForCount}
\alias{getCutDateForCount}
\title{Get Date at Which an Event Count is Reached}
\usage{
getCutDateForCount(x, count)
}
\arguments{
\item{x}{a time-to-event dataset, e.g., generated by \code{simPWSurv}}

\item{count}{event count at which dataset is to be cut off for analysis}
}
\value{
The a numeric value with the \code{cte} from the input dataset at which the targeted event count
is reached, or if the final event count is never reached, the final \code{cte} at which an event occurs.
}
\description{
Get Date at Which an Event Count is Reached
}
\examples{
library(dplyr)
# Use default enrollment and calendar cut date for 50 events in Positive stratum
x <- simPWSurv(n=200,
               strata = tibble::tibble(Stratum=c("Positive","Negative"), p = c(.5, .5)),
               failRates = tibble::tibble(Stratum = rep(c("Positive","Negative"),2),
                                          period = rep(1, 4),
                                          Treatment = c(rep("Control", 2), 
                                                        rep("Experimental", 2)),
                                          duration = rep(1, 4),
                                          rate = log(2) / c(6, 9, 9, 12)
                                          ),
               dropoutRates = tibble::tibble(Stratum = rep(c("Positive","Negative"),2),
                                             period = rep(1, 4),
                                             Treatment = c(rep("Control", 2), 
                                                           rep("Experimental", 2)),
                                             duration = rep(1, 4),
                                             rate = rep(.001, 4)
                                            )
              )
d <- getCutDateForCount(filter(x,Stratum=="Positive"),count=50)
y <- cutData(x,cutDate=d)
table(y$Stratum,y$event)

}
