% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlr_weight.R
\name{mb}
\alias{mb}
\title{Magirr and Burman weighting function}
\usage{
mb(delay = 4, w_max = Inf)
}
\arguments{
\item{delay}{The initial delay period where weights increase;
after this, weights are constant at the final weight in the delay period.}

\item{w_max}{Maximum weight to be returned.
Set \code{delay = Inf}, \code{w_max = 2} to be consistent with recommendation of
Magirr (2021).}
}
\value{
A list of parameters of the Magirr and Burman weighting function
}
\description{
Magirr and Burman weighting function
}
\examples{
mb(delay = 6, w_max = 2)
}
