% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb_weight.R
\name{mb_weight}
\alias{mb_weight}
\title{Magirr and Burman modestly weighted logrank tests}
\usage{
mb_weight(x, delay = 4, w_max = Inf)
}
\arguments{
\item{x}{A \code{\link[=counting_process]{counting_process()}}-class data frame with a counting process
dataset.}

\item{delay}{The initial delay period where weights increase;
after this, weights are constant at the final weight in the delay period.}

\item{w_max}{Maximum weight to be returned.
Set \code{delay = Inf}, \code{w_max = 2} to be consistent with recommendation of
Magirr (2021).}
}
\value{
A list.
}
\description{
Computes Magirr-Burman weights and adds them to a dataset created by
\code{\link[=counting_process]{counting_process()}}.
These weights can then be used to compute a z-statistic for the
modestly weighted logrank test proposed.
}
