% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestone.R
\name{milestone}
\alias{milestone}
\title{Milestone test for two survival curves}
\usage{
milestone(data, ms_time)
}
\arguments{
\item{data}{Data frame containing at least 3 columns:
\itemize{
\item \code{tte} - Time to event.
\item \code{event} - Event indicator.
\item \code{treatment} - Grouping variable.
}}

\item{ms_time}{Milestone analysis time.}
}
\value{
A data frame containing the method (\code{method}), always "milestone",
test statistics (\code{z}), milestone time point (\code{ms_time}),
survival rate of the control arm (\code{surv0}),
survival rate of the experimental arm (\code{surv1}),
survival difference between the experimental and control arm (\code{surv_diff}),
standard error of the control arm (\code{std_err0}),
standard error of the experimental arm (\code{std_err1}).
}
\description{
Milestone test for two survival curves
}
\examples{
sim_pw_surv(n = 200) |>
  cut_data_by_event(150) |>
  milestone(10)
}
