% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rpwexp_inverse_cdf_cpp}
\alias{rpwexp_inverse_cdf_cpp}
\title{Piecewise exponential distribution using the inverse CDF method}
\usage{
rpwexp_inverse_cdf_cpp(n, fail_rate)
}
\arguments{
\item{n}{Number of observations to be generated.}

\item{fail_rate}{A data frame containing \code{duration} and \code{rate} variables.}
}
\description{
Piecewise exponential distribution using the inverse CDF method
}
