% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rpwexp_naive_cpp}
\alias{rpwexp_naive_cpp}
\title{Piecewise exponential distribution using the naive method}
\usage{
rpwexp_naive_cpp(n, fail_rate)
}
\arguments{
\item{n}{Number of observations to be generated.}

\item{fail_rate}{A data frame containing \code{duration} and \code{rate} variables.}
}
\description{
Piecewise exponential distribution using the naive method
}
