% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenFH.r
\name{tenFH}
\alias{tenFH}
\title{Fleming-Harrington Weighted Logrank Tests}
\usage{
tenFH(
  x = simPWSurv(n = 200) \%>\% cutDataAtCount(150) \%>\% tensurv(txval =
    "Experimental"),
  rg = tibble(rho = c(0, 0, 1, 1), gamma = c(0, 1, 0, 1)),
  returnVariance = FALSE
)
}
\arguments{
\item{x}{a \code{tensurv}-class \code{tibble} with a counting process dataset}

\item{rg}{a \code{tibble} with variables \code{rho} and \code{gamma}, both greater than equal
to zero, to specify one Fleming-Harrington weighted logrank test per row;
Default: tibble(rho = c(0, 0, 1, 1), gamma = c(0, 1, 0, 1))}

\item{returnVariance}{a logical flag that, if true, adds columns
estimated variance for weighted sum of observed minus expected; see details; Default: FALSE}
}
\value{
a `tibble` with \code{rg} as input and the FH test statistic
for the data in \code{x}
(\code{Z}, a directional square root of the usual weighted logrank test);
if variance calculations are specified (e.g., to be used for covariances in a combination test),
the this will be returned in the column \code{Var}
}
\description{
With output from the function \code{tensurv}
}
\details{
DETAILS
}
\examples{
library(tidyr)
# Use default enrollment and event rates at cut at 100 events
x <- simPWSurv(n=200) \%>\% cutDataAtCount(100) \%>\% tensurv(txval="Experimental")
# compute logrank (FH(0,0)) and FH(0,1)
tenFH(x,rg=tibble(rho=c(0,0),gamma=c(0,1)))
}
