% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wMB.R
\name{wMB}
\alias{wMB}
\title{Magirr and Burman Modestly Weighted Logrank Tests}
\usage{
wMB(x, delay = 4)
}
\arguments{
\item{x}{a \code{tensurv}-class \code{tibble} with a counting process dataset}

\item{delay}{The initial delay period where weights increase; 
after this, weights are constant at the final weigh in the delay period}
}
\value{
a vector with weights for the Magirr-Burman weighted logrank test
for the data in \code{x}
}
\description{
Magirr and Burman (2019) have proposed a weighted logrank test to have better power than
the logrank test when the treatment effect is delayed, but to still maintain good power under 
a proportional hazards assumption. 
The weights for some early interval specified by the user are the inverse of the combined treatment group
empirical survival distribution; see details. 
After this initial period, weights are constant at the maximum of the previous weights.
Another advantage of the test is that under strong null hypothesis that the underlying survival in the control group
is greater than or equal to underlying survival in the experimental group, 
Type I error is controlled as the specified level.
}
\details{
This function computes Magirr-Burman weights and adds them to a dataset created by the \code{tensurv()} function.
These weights can then be used to compute a Z-statistic for the modestly weighted logrank test proposed.


We define \eqn{t^*} to be the input variable \code{delay}.
This specifies an initial period during which weights increase.
To define specific weights, we let \eqn{S(t)} denote the Kaplan-Meier survival estimate at time \eqn{t}
for the combined data (control plus experimental treatment groups).
The weight at time \eqn{t} is then defined as
\deqn{w(t)=S(\min(t,t^*))^{-1}.}
}
\examples{
library(tidyr)
library(dplyr)
# Use default enrollment and event rates at cut at 100 events
x <- simPWSurv(n=200) \%>\% cutDataAtCount(125) \%>\% tensurv(txval="Experimental")
# compute Magirr-Burman weights with 
ZMB <- x \%>\% wMB(6) \%>\% 
             summarize(S=sum(OminusE*wMB),V=sum(Var*wMB^2),Z=S/sqrt(V))
# Compute p-value of modestly weighted logrank of Magirr-Burman
pnorm(ZMB$Z)
}
\references{
Magirr, Dominic, and Carl‐Fredrik Burman. 
"Modestly weighted logrank tests." 
\emph{Statistics in Medicine} 38.20 (2019): 3782-3790.
}
