% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutData.R
\name{cut_data_by_date}
\alias{cut_data_by_date}
\title{Cut a Dataset for Analysis at a Specified Date}
\usage{
cut_data_by_date(x, cutDate)
}
\arguments{
\item{x}{a time-to-event dataset, e.g., generated by \code{simPWSurv}}

\item{cutDate}{date relative to start of randomization (\code{cte} from input dataset)
at which dataset is to be cut off for analysis}
}
\value{
A dataset ready for survival analysis
}
\description{
Cut a Dataset for Analysis at a Specified Date
}
\examples{
# Use default enrollment and event rates and
# cut at calendar time 5 after start of randomization
library(dplyr)
simPWSurv(n = 20) \%>\% cut_data_by_date(5)

}
