% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutDataAtCount.R
\name{cutDataAtCount}
\alias{cutDataAtCount}
\title{Cut a Dataset for Analysis at a Specified Event Count}
\usage{
cutDataAtCount(x, count)
}
\arguments{
\item{x}{a time-to-event dataset, e.g., generated by \code{simPWSurv}}

\item{count}{event count at which data cutoff is to be made}
}
\value{
a \code{tibble} ready for survival analysis, including culumns time to event (\code{tte}),
\code{event}, the \code{stratum} and the \code{treatment.}
}
\description{
\code{cutDataAtCount} takes a time-to-event data set and cuts the data at which an
event count is reached.
}
\examples{
library(tidyr)
# Use default enrollment and event rates at cut at 100 events
x <- simPWSurv(n=200) \%>\% cutDataAtCount(100)
table(x$event,x$Treatment)

}
