% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlr_weight.R
\name{early_zero}
\alias{early_zero}
\title{Zero early weighting function}
\usage{
early_zero(early_period)
}
\arguments{
\item{early_period}{The initial delay period where weights increase;
after this, weights are constant at the final weight in the delay period.}
}
\value{
A list of parameters of the zero early weighting function
}
\description{
Zero early weighting function
}
\examples{
early_zero(6)
}
\references{
Xu, Z., Zhen, B., Park, Y., & Zhu, B. (2017).
"Designing therapeutic cancer vaccine trials with delayed treatment effect."
}
