% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/early_zero_weight.R
\name{early_zero_weight}
\alias{early_zero_weight}
\title{Zero early weight for weighted logrank tests}
\usage{
early_zero_weight(x, early_period = 4, fail_rate = NULL)
}
\arguments{
\item{x}{A \code{\link[=counting_process]{counting_process()}}-class data frame with a counting process dataset.}

\item{early_period}{The initial delay period where weights increase;
after this, weights are constant at the final weight in the delay period.}

\item{fail_rate}{A data frame record the failure rate.}
}
\value{
A data frame. The column \code{weight} contains the weights for the
early zero weighted logrank test for the data in \code{x}.
}
\description{
Zero early weight for weighted logrank tests
}
