% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCutDateForCount.R
\name{get_cut_date_by_event}
\alias{get_cut_date_by_event}
\title{Get Date at Which an Event Count is Reached}
\usage{
get_cut_date_by_event(x, count)
}
\arguments{
\item{x}{a time-to-event dataset, e.g., generated by \code{sim_pw_surv}}

\item{count}{event count at which dataset is to be cut off for analysis}
}
\value{
The a numeric value with the \code{cte} from the input dataset at which the targeted event count
is reached, or if the final event count is never reached, the final \code{cte} at which an event occurs.
}
\description{
Get Date at Which an Event Count is Reached
}
\examples{
library(dplyr)
library(tibble)

# Use default enrollment and calendar cut date for 50 events in Positive stratum
x <- sim_pw_surv(
  n = 200,
  stratum = tibble(stratum = c("Positive", "Negative"),
                  p = c(.5, .5)),
  fail_rate = tibble(stratum = rep(c("Positive","Negative"), 2),
                     period = rep(1, 4),
                     treatment = c(rep("control", 2), rep("experimental", 2)),
                     duration = rep(1, 4),
                     rate = log(2) / c(6, 9, 9, 12)),
  dropout_rate = tibble(stratum = rep(c("Positive", "Negative"),2),
                        period = rep(1, 4),
                        treatment = c(rep("control", 2), rep("experimental", 2)),
                        duration = rep(1, 4),
                        rate = rep(.001, 4)))

d <- get_cut_date_by_event(x \%>\% filter(stratum == "Positive"), count = 50)

y <- cut_data_by_date(x, cut_date = d)
table(y$stratum, y$event)

}
