% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxcombo.R
\name{maxcombo}
\alias{maxcombo}
\title{Maxcombo test}
\usage{
maxcombo(data, test1, test2, ...)
}
\arguments{
\item{data}{a tte dataset}

\item{test1}{maxcombo test1}

\item{test2}{maxcombo test2}

\item{...}{additional tests}
}
\value{
pvalues
}
\description{
Maxcombo test
}
\examples{
sim_pw_surv(n = 200) |>
  cut_data_by_event(150) |>
  maxcombo(test1 = wlr(data, rho = 0, gamma = 0) |> quote(),
           test2 = wlr(data, rho = 0, gamma = 0.5) |> quote())
}
