% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxcombo.R
\name{maxcombo}
\alias{maxcombo}
\title{Maxcombo test}
\usage{
maxcombo(
  data = cut_data_by_event(sim_pw_surv(n = 200), 150),
  rho = c(0, 0, 1),
  gamma = c(0, 1, 1),
  return_variance = FALSE,
  return_corr = FALSE
)
}
\arguments{
\item{data}{a tte dataset}

\item{rho}{Numeric vector passed to \code{\link[=fh_weight]{fh_weight()}}. Must be greater
than or equal to zero. Must be the same length as \code{gamma}.}

\item{gamma}{Numeric vector passed to \code{\link[=fh_weight]{fh_weight()}}. Must be
greater than or equal to zero. Must be the same length as \code{rho}.}
}
\value{
pvalues
}
\description{
WARNING: This experimental function is a work-in-progress. The function
arguments will change as we add additional features.
}
\examples{
sim_pw_surv(n = 200) |>
  cut_data_by_event(150) |>
  maxcombo(rho = c(0, 0), gamma = c(0, 1))
}
\seealso{
\code{\link[=fh_weight]{fh_weight()}}
}
