% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestone.R
\name{milestone}
\alias{milestone}
\title{Milestone test for two survival curves}
\usage{
milestone(data, ms_time)
}
\arguments{
\item{data}{Dataset contains at least 3 columns: \code{tte} (time to event),
\code{event} (event indicator), and \code{treatment} (grouping variable).}

\item{ms_time}{Milestone analysis time}
}
\value{
A data frame containing the test statistics
}
\description{
Milestone test for two survival curves
}
\examples{
sim_pw_surv(n = 200) |>
  cut_data_by_event(150) |>
  milestone(10)
}
