% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue_maxcombo.R
\name{pvalue_maxcombo}
\alias{pvalue_maxcombo}
\title{MaxCombo p-value}
\usage{
pvalue_maxcombo(
  z,
  dummy_var,
  algorithm = GenzBretz(maxpts = 50000, abseps = 1e-05)
)
}
\arguments{
\item{z}{a dataset output from \code{tenFHcorr()}; see examples.}

\item{dummy_var}{a dummy input that allows \code{group_map()} to be used to
compute p-values for multiple simulations.}

\item{algorithm}{This is passed directly to the \code{algorithm} argument in the \code{mvtnorm::pmvnorm()}}
}
\value{
A numeric p-value
}
\description{
\code{pvalue_maxcombo()} computes p-values for the MaxCombo test
based on output from \code{simtrial::tenFHcorr()}.
This is still in an experimental stage and is intended for use with
the \code{simtrial::sim_fixed_n()} trial simulation routine.
However, it can also be used to analyze clinical trial data such as that provided in the
ADaM ADTTE format.
}
\examples{
library(tidyr)
library(tibble)
library(dplyr)

# example 1
x <- sim_fixed_n(nsim = 1,
            timing_type = 5,
            rho_gamma = tibble(rho = c(0, 0, 1),
                        gamma = c(0, 1, 1)))
head(x)
pvalue_maxcombo(x)

# example 2
# Only use cuts for events, events + min follow-up
xx <- sim_fixed_n(nsim = 100,
             timing_type = 5,
             rho_gamma = tibble(rho = c(0, 0, 1),
                         gamma = c(0, 1, 1)))
head(xx)
# MaxCombo power estimate for cutoff at max of targeted events, minimum follow-up
p <- xx \%>\% group_by(Sim) \%>\% group_map(pvalue_maxcombo) \%>\% unlist()
mean(p < .025)

}
