% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmst.R
\name{rmst}
\alias{rmst}
\title{RMST difference of 2 arms}
\usage{
rmst(
  data,
  tau = 10,
  var_label_tte = "tte",
  var_label_event = "event",
  var_label_group = "treatment",
  reference = "control",
  alpha = 0.05
)
}
\arguments{
\item{data}{A time-to-event dataset with a column \code{tte} indicating the
survival time and a column of \code{event} indicating whether it is
event or censor.}

\item{tau}{RMST analysis time.}

\item{var_label_tte}{Column name of the TTE variable.}

\item{var_label_event}{Column name of the event variable.}

\item{var_label_group}{Column name of the grouping variable.}

\item{reference}{A group label indicating the reference group.}

\item{alpha}{Type I error.}
}
\value{
The z statistics.
}
\description{
RMST difference of 2 arms
}
\examples{
data(ex1_delayed_effect)
rmst(
  data = ex1_delayed_effect,
  var_label_tte = "month",
  var_label_event = "evntd",
  var_label_group = "trt",
  tau = 10,
  reference = "0"
)
}
