% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmst.R
\name{rmst}
\alias{rmst}
\title{Title}
\usage{
rmst(
  data,
  tau = 10,
  var_label_tte = "tte",
  var_label_event = "event",
  var_label_group = "treatment",
  reference = "control",
  alpha = 0.05
)
}
\arguments{
\item{data}{a time -to-event dataset with a column \code{tte} indicating the survival time and
a column of \code{event} indicating whether it is event or censor.}

\item{tau}{RMST analysis time}

\item{reference}{a group label indicating the reference group}

\item{alpha}{type I error}
}
\value{
the z statistics
}
\description{
Title
}
\examples{
data(ex1_delayed_effect)
rmst(data = ex1_delayed_effect,
     var_label_tte = "month",
     var_label_event = "evntd",
     var_label_group = "trt",
     tau = 10,
     reference = "0")
}
