% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rpwexpinvRcpp}
\alias{rpwexpinvRcpp}
\title{The Piecewise Exponential Distribution using inverse CDF method in C++}
\usage{
rpwexpinvRcpp(n, fail_rate)
}
\arguments{
\item{n}{Number of observations to be generated.}

\item{fail_rate}{A dataframe containing \code{duration} and \code{rate} variables.}
}
\description{
The Piecewise Exponential Distribution using inverse CDF method in C++
}
