% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rpwexpRcpp}
\alias{rpwexpRcpp}
\title{The Piecewise Exponential Distribution in C++}
\usage{
rpwexpRcpp(n, failRates)
}
\arguments{
\item{n}{Number of observations to be generated.}

\item{failRates}{A dataframe containing \code{duration} and \code{rate} variables.}
}
\description{
The Piecewise Exponential Distribution in C++
}
