% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rpwexpRcpp}
\alias{rpwexpRcpp}
\title{The piecewise exponential distribution in C++}
\usage{
rpwexpRcpp(n, fail_rate)
}
\arguments{
\item{n}{Number of observations to be generated.}

\item{fail_rate}{A data frame containing \code{duration} and \code{rate} variables.}
}
\description{
The piecewise exponential distribution in C++
}
