% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simtrial-package.R
\docType{package}
\name{simtrial-package}
\alias{simtrial}
\alias{simtrial-package}
\title{simtrial: Clinical Trial Simulation}
\description{
simtrial provides some basic routines for simulating a clinical trial. The primary intent is to provide some tools to generate trial simulations for trials with time to event outcomes. Piecewise exponential failure rates and piecewise constant enrollment rates are the underlying mechanism used to simulate a broad range of scenarios. However, the basic generation of data is done using pipes to allow maximum flexibility for users to meet different needs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://merck.github.io/simtrial/}
  \item \url{https://github.com/Merck/simtrial}
  \item Report bugs at \url{https://github.com/Merck/simtrial/issues}
}

}
\author{
\strong{Maintainer}: Yujie Zhao \email{yujie.zhao@merck.com} [contributor]

Authors:
\itemize{
  \item Keaven Anderson \email{keaven_anderson@merck.com}
  \item Yilong Zhang \email{elong0527@gmail.com}
}

Other contributors:
\itemize{
  \item Nan Xiao [contributor]
  \item Jianxiao Yang [contributor]
  \item Lili Ling [contributor]
  \item Xintong Li [contributor]
  \item Ruixue Wang [contributor]
  \item Yi Cui [contributor]
  \item Ping Yang [contributor]
  \item Yalin Zhu [contributor]
  \item Heng Zhou [contributor]
  \item Amin Shirazi [contributor]
  \item Merck & Co., Inc., Rahway, NJ, USA and its affiliates [copyright holder]
}

}
\keyword{internal}
