% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simtrial-package.R
\docType{package}
\name{simtrial-package}
\alias{simtrial}
\alias{simtrial-package}
\title{simtrial: Clinical Trial Simulation}
\description{
simtrial provides some basic routines for simulating a clinical trial. 
    The primary intent is to provide some tools to generate trial simulations for trials with time to event outcomes. 
    Piecewise exponential failure rates and piecewise constant enrollment rates are the underlying mechanism used 
    to simulate a broad range of scenarios.
    However, the basic generation of data is done using pipes to allow maximum flexibility for users to meet different needs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://merck.github.io/simtrial/}
  \item \url{https://github.com/Merck/simtrial}
  \item Report bugs at \url{https://github.com/Merck/simtrial/issues}
}

}
\author{
\strong{Maintainer}: Yilong Zhang \email{yilong.zhang@merck.com}

Authors:
\itemize{
  \item Keaven Anderson \email{keaven_anderson@merck.com}
}

Other contributors:
\itemize{
  \item Amin Shirazi \email{ashirazist@gmail.com} [contributor]
  \item Ruixue Wang \email{ruixue.wang@merck.com} [contributor]
  \item Yi Cui \email{yi.cui@merck.com} [contributor]
  \item Ping Yang \email{ping.yang1@merck.com} [contributor]
  \item Yalin Zhu \email{yalin.zhu@merck.com} [contributor]
  \item Heng Zhou \email{heng.zhou@merck.com} [contributor]
  \item Merck Sharp & Dohme Corp [copyright holder]
}

}
\keyword{internal}
