% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlr.R
\name{wlr}
\alias{wlr}
\title{Weighted logrank test}
\usage{
wlr(data, weight)
}
\arguments{
\item{data}{cutted dataset generated by sim_pw_surv}

\item{weight}{weighting functions, such as \code{fh_weight}, \code{mb_weight}, and \code{early_zero_weight}.}
}
\value{
test results
}
\description{
Weighted logrank test
}
\examples{
sim_pw_surv(n = 200) |>
  cut_data_by_event(150) |>
  wlr(weight = fh(rho = 0, gamma = 0))

sim_pw_surv(n = 200) |>
  cut_data_by_event(150) |>
  wlr(weight = mb(delay = 4, w_max = 2))

sim_pw_surv(n = 200) |>
  cut_data_by_event(150) |>
  wlr(weight = early_zero(early_period = 4))

}
