% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlr.R
\name{wlr}
\alias{wlr}
\title{Weighted logrank test}
\usage{
wlr(data, weight, return_variance = FALSE)
}
\arguments{
\item{data}{cutted dataset generated by sim_pw_surv}

\item{weight}{weighting functions, such as \code{\link[=fh_weight]{fh_weight()}}, \code{\link[=mb_weight]{mb_weight()}}, and
\code{\link[=early_zero_weight]{early_zero_weight()}}.}

\item{return_variance}{A logical flag that, if \code{TRUE}, adds columns
estimated variance for weighted sum of observed minus expected;
see details; Default: \code{FALSE}.}
}
\value{
test results
}
\description{
Weighted logrank test
}
\examples{
x <- sim_pw_surv(n = 200) |> cut_data_by_event(100)

# Example 1: WLR test with FH wights
x |> wlr(weight = fh(rho = 0, gamma = 1))
x |> wlr(weight = fh(rho = 0, gamma = 1), return_variance = TRUE)

# Example 2: WLR test with MB wights
x |> wlr(weight = mb(delay = 4, w_max = 2))

# Example 3: WLR test with early zero wights
x |> wlr(weight = early_zero(early_period = 4))
}
